cambios de carpetas y arreglos en proyecto react
