+++ softech_project/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "softech_project",
+  "name": "softech_proyect",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
