+++ softech_project/src/App.js
@@ -1,25 +1,28 @@
-import logo from './logo.svg';
-import './App.css';
 
-function App() {
+import './App.css';
+import Sidebar from './features/sidebar/Sidebar';
+import Navbar from './features/navbar/Navbar';
+import { useState } from 'react';
+
+const App =()=> {
+  const [contadorValue,updaterContador]=useState(0);
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+   <div class="template">
+
+    <Sidebar contadorValue={contadorValue} updaterContador={updaterContador} />
+    <main>
+      <Navbar contadorValue={contadorValue} />
+    
+        <section class="contenido">
+
+
+        </section>
+    </main>
+   </div> 
+    
   );
 }
 
 export default App;
+
+
