+++ softech_project/src/features/sidebar/logoSporty.png
